feat: execute persistent live targets and rules
Persist and reconcile target and reducing-risk intent across partial fills and restarts while exposing the shared strategy facade.

Closes ml4t/live#14.

+++ src/ml4t/live/cli/main.py
@@ -148,7 +148,7 @@ class IntentPrintingSafeBroker(SafeBroker):
     async def submit_order_async(
         self,
         asset: str,
-        quantity: int,
+        quantity: float,
         side: OrderSide | None = None,
         order_type: OrderType = OrderType.MARKET,
         limit_price: float | None = None,
